Send extracted thumbnails data directly to stdout
The data is no longer base64 encoded. This saves time since we don't
need to start a Python ZeroVM job for this. It also saves some
bandwidth since base64 encoding expands the input by 33%.

This change breaks compatibility with IE 9 since we now require
support for XMLHttpRequest Level 2:

  http://caniuse.com/#search=xhr2
  http://www.w3.org/TR/XMLHttpRequest2/

Some 81% of the users have this support, and for a bleeding-edge
project like ours, I believe we can require a current browser.